+++ corehq/apps/callcenter/tests/__init__.py
@@ -1,5 +1,6 @@
 try:
     from corehq.apps.callcenter.tests.test_indicators import *
+    from corehq.apps.callcenter.tests.test_indicator_fixture import *
 except ImportError, e:
     # for some reason the test harness squashes these so log them here for clarity
     # otherwise debugging is a pain
