--- corehq/apps/reportfixtures/models.py
@@ -1,5 +1,0 @@
-from couchdbkit.ext.django.schema import Document
-
-
-class _(Document):
-    pass
